backend/requirements.txt
django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-two-factor-auth
django[argon2]
django-fernet-fields==0.6.*
django-ratelimit==4.1.*
django-cors-headers==4.3.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
phonenumbers==8.13.*
django-otp==1.2.*
africastalking